- Register OGG Vorbis codec - Rename namespace
